feat: make abstract signer
