+++ packages/network/src/signer/signers/index.ts
@@ -1,3 +1,4 @@
 export * from './utils';
 export * from './vechain-private-key-signer';
+export * from './vechain-abstract-signer';
 export * from './types.d';
