+++ packages/network/src/signer/signers/vechain-abstract-signer/index.ts
@@ -1,0 +1,1 @@
+export * from './vechain-abstract-signer';
